+++ App/Navigation/index.js
@@ -1,0 +1,5 @@
+import CollectionStack from './CollectionStack';
+
+export {
+    CollectionStack,
+}
